convert unit tests to integration tests
