+++ prime-router/src/test/kotlin/serializers/Hl7SerializerTests.kt
@@ -67,105 +67,6 @@ class Hl7SerializerTests {
     private val emptyTerser = Terser(ORU_R01())
 
     @Test
-    fun `test checkLIVDValueExists`() {
-        val settings = FileSettings("./settings")
-        val serializer = Hl7Serializer(UnitTestUtils.simpleMetadata, settings)
-
-        // happy path success
-        val modelCOVIDSeqTest = serializer.checkLIVDValueExists("Model", "COVIDSeq Test")
-        assertThat(modelCOVIDSeqTest).isTrue()
-
-        // happy path fail
-        val modelFake = serializer.checkLIVDValueExists("Model", "Fake")
-        assertThat(modelFake).isFalse()
-
-        // unhappy path input 1
-        val fakeCOVIDSeqTest = serializer.checkLIVDValueExists("fake", "COVIDSeq Test")
-        assertThat(fakeCOVIDSeqTest).isFalse()
-
-        // unhappy path input 1 and 2
-        val fakeFake = serializer.checkLIVDValueExists("fake", "fake")
-        assertThat(fakeFake).isFalse()
-    }
-
-    @Test
-    fun `test confirmObservationOrder`() {
-        val settings = FileSettings("./settings")
-        val serializer = Hl7Serializer(UnitTestUtils.simpleMetadata, settings)
-
-        // message where first OBX segment contains the test result
-        val sampleRegMessage = """MSH|^~\&|CDC PRIME - Atlanta, Georgia (Dekalb)^2.16.840.1.114222.4.1.237821^ISO|Avante at Ormond Beach^10D0876999^CLIA|||20210210170737||ORU^R01^ORU_R01|371784|P|2.5.1|||NE|NE|USA||||PHLabReportNoAck^ELR_Receiver^2.16.840.1.113883.9.11^ISO
-SFT|Centers for Disease Control and Prevention|0.1-SNAPSHOT|PRIME ReportStream|0.1-SNAPSHOT||20210210
-PID|1||2a14112c-ece1-4f82-915c-7b3a8d152eda^^^Avante at Ormond Beach^PI||Test^Kareem^Millie^^^^L||19580810|F||2106-3^White^HL70005^^^^2.5.1|688 Leighann Inlet^^South Rodneychester^TX^67071||^PRN^^roscoe.wilkinson@email.com^1^211^2240784|||||||||U^Unknown^HL70189||||||||N
-ORC|RE|73a6e9bd-aaec-418e-813a-0ad33366ca85|73a6e9bd-aaec-418e-813a-0ad33366ca85|||||||||1629082607^Eddin^Husam^^^^^^CMS&2.16.840.1.113883.3.249&ISO^^^^NPI||^WPN^^^1^386^6825220|20210209||||||Avante at Ormond Beach|170 North King Road^^Ormond Beach^FL^32174^^^^12127|^WPN^^jbrush@avantecenters.com^1^407^7397506|^^^^32174
-OBR|1|73a6e9bd-aaec-418e-813a-0ad33366ca85||94558-4^SARS-CoV-2 (COVID-19) Ag [Presence] in Respiratory specimen by Rapid immunoassay^LN|||202102090000-0600|202102090000-0600||||||||1629082607^Eddin^Husam^^^^^^CMS&2.16.840.1.113883.3.249&ISO^^^^NPI|^WPN^^^1^386^6825220|||||202102090000-0600|||F
-OBX|1|CWE|94558-4^SARS-CoV-2 (COVID-19) Ag [Presence] in Respiratory specimen by Rapid immunoassay^LN||260415000^Not detected^SCT|||N^Normal (applies to non-numeric results)^HL70078|||F|||202102090000-0600|||CareStart COVID-19 Antigen test_Access Bio, Inc._EUA^^99ELR||202102090000-0600||||Avante at Ormond Beach^^^^^CLIA&2.16.840.1.113883.4.7&ISO^^^^10D0876999^CLIA|170 North King Road^^Ormond Beach^FL^32174^^^^12127
-NTE|1|L|This is a comment|RE
-OBX|2|CWE|95418-0^Whether patient is employed in a healthcare setting^LN^^^^2.69||Y^Yes^HL70136||||||F|||202102090000-0600|||||||||||||||QST
-OBX|3|CWE|95417-2^First test for condition of interest^LN^^^^2.69||Y^Yes^HL70136||||||F|||202102090000-0600|||||||||||||||QST
-OBX|4|CWE|95421-4^Resides in a congregate care setting^LN^^^^2.69||N^No^HL70136||||||F|||202102090000-0600|||||||||||||||QST
-OBX|5|CWE|95419-8^Has symptoms related to condition of interest^LN^^^^2.69||N^No^HL70136||||||F|||202102090000-0600|||||||||||||||QST
-SPM|1|||258500001^Nasopharyngeal swab^SCT||||71836000^Nasopharyngeal structure (body structure)^SCT^^^^2020-09-01|||||||||202102090000-0600^202102090000-0600"""
-
-        arrangeTest(sampleRegMessage).run {
-            val oruR01SampleReg: ORU_R01? = this as? ORU_R01
-            // save OBX-3 for first OBX segment prior to the function run
-            val sampleObsID = oruR01SampleReg
-                ?.patienT_RESULT?.ordeR_OBSERVATION?.observation?.obx?.observationIdentifier?.identifier
-
-            // assert
-            serializer.organizeObservationOrder(this).run {
-                val happyORUR01 = this as? ORU_R01
-                // extract OBX-3 for first OBX segment after function run
-                val happyFullObsIdentifier = happyORUR01
-                    ?.patienT_RESULT?.ordeR_OBSERVATION?.observation?.obx?.observationIdentifier?.identifier
-                // since first OBX segment does contain the test result, the sample and final OBX-3 values should match
-                assertThat(happyFullObsIdentifier).isEqualTo(sampleObsID)
-            }
-        }
-        // message where first OBX segment does not contain the test result
-        val sampleComplexMessage = """MSH|^~\&|CDC PRIME - Atlanta, Georgia (Dekalb)^2.16.840.1.114222.4.1.237821^ISO|Avante at Ormond Beach^10D0876999^CLIA|||20210210170737||ORU^R01^ORU_R01|371784|P|2.5.1|||NE|NE|USA||||PHLabReportNoAck^ELR_Receiver^2.16.840.1.113883.9.11^ISO
- SFT|Centers for Disease Control and Prevention|0.1-SNAPSHOT|PRIME ReportStream|0.1-SNAPSHOT||20210210
- PID|1||2a14112c-ece1-4f82-915c-7b3a8d152eda^^^Avante at Ormond Beach^PI||Test^Kareem^Millie^^^^L||19580810|F||2106-3^White^HL70005^^^^2.5.1|688 Leighann Inlet^^South Rodneychester^TX^67071||^PRN^^roscoe.wilkinson@email.com^1^211^2240784|||||||||U^Unknown^HL70189||||||||N
- ORC|RE|73a6e9bd-aaec-418e-813a-0ad33366ca85|73a6e9bd-aaec-418e-813a-0ad33366ca85|||||||||1629082607^Eddin^Husam^^^^^^CMS&2.16.840.1.113883.3.249&ISO^^^^NPI||^WPN^^^1^386^6825220|20210209||||||Avante at Ormond Beach|170 North King Road^^Ormond Beach^FL^32174^^^^12127|^WPN^^jbrush@avantecenters.com^1^407^7397506|^^^^32174
- OBR|1|73a6e9bd-aaec-418e-813a-0ad33366ca85||94558-4^SARS-CoV-2 (COVID-19) Ag [Presence] in Respiratory specimen by Rapid immunoassay^LN|||202102090000-0600|202102090000-0600||||||||1629082607^Eddin^Husam^^^^^^CMS&2.16.840.1.113883.3.249&ISO^^^^NPI|^WPN^^^1^386^6825220|||||202102090000-0600|||F
- OBX|1|CWE|95418-0^Whether patient is employed in a healthcare setting^LN^^^^2.69||Y^Yes^HL70136||||||F|||202102090000-0600|||||||||||||||QST
- OBX|2|CWE|95417-2^First test for condition of interest^LN^^^^2.69||Y^Yes^HL70136||||||F|||202102090000-0600|||||||||||||||QST
- OBX|3|CWE|95421-4^Resides in a congregate care setting^LN^^^^2.69||N^No^HL70136||||||F|||202102090000-0600|||||||||||||||QST
- OBX|4|CWE|95419-8^Has symptoms related to condition of interest^LN^^^^2.69||N^No^HL70136||||||F|||202102090000-0600|||||||||||||||QST
- OBX|5|CWE|94558-4^SARS-CoV-2 (COVID-19) Ag [Presence] in Respiratory specimen by Rapid immunoassay^LN||260415000^Not detected^SCT|||N^Normal (applies to non-numeric results)^HL70078|||F|||202102090000-0600|||CareStart COVID-19 Antigen test_Access Bio, Inc._EUA^^99ELR||202102090000-0600||||Avante at Ormond Beach^^^^^CLIA&2.16.840.1.113883.4.7&ISO^^^^10D0876999^CLIA|170 North King Road^^Ormond Beach^FL^32174^^^^12127
- NTE|1|L|This is a comment|RE
- SPM|1|||258500001^Nasopharyngeal swab^SCT||||71836000^Nasopharyngeal structure (body structure)^SCT^^^^2020-09-01|||||||||202102090000-0600^202102090000-0600"""
-
-        arrangeTest(sampleComplexMessage).run {
-            val oruR01SampleComplex: ORU_R01? = this as? ORU_R01
-            // save OBX-3 for first OBX segment prior to the function run
-            val sampleComplexObsID = oruR01SampleComplex
-                ?.patienT_RESULT?.ordeR_OBSERVATION?.observation?.obx?.observationIdentifier?.identifier
-
-            // assert
-            serializer.organizeObservationOrder(this).run {
-                val happyComplexORUR01 = this as? ORU_R01
-                // extract OBX-3 for first OBX segment after function run
-                val happyComplexObsIdentifier = happyComplexORUR01
-                    ?.patienT_RESULT?.ordeR_OBSERVATION?.observation?.obx?.observationIdentifier?.identifier
-                // since first OBX segment does not contain the test result, the sample and final OBX-3 values
-                // should not match
-                assertThat(happyComplexObsIdentifier).isNotEqualTo(sampleComplexObsID)
-
-                // the OBX segments were rearranged. But the OBX set IDs should still be sequential
-                val happyComplexSetID = happyComplexORUR01
-                    ?.patienT_RESULT?.ordeR_OBSERVATION?.observation?.obx?.setIDOBX
-                assertThat(happyComplexSetID.toString()).isEqualTo("1")
-
-                // the OBX segment with the test result has an attached NTE. That should be moved along with the OBX
-                val happyComplexNTE = happyComplexORUR01?.patienT_RESULT?.ordeR_OBSERVATION?.observation?.nte
-                assertThat(happyComplexNTE).isNotNull()
-            }
-        }
-    }
-
-    @Test
     fun `test XTN phone decoding`() {
         val settings = FileSettings("./settings")
         val serializer = Hl7Serializer(UnitTestUtils.simpleMetadata, settings)
